Small styling fixes - moved header back to components, table header in orange area, breakpoitn for table sizing
