+++ frontend/src/grant-info/components/GrantPage.tsx
@@ -1,5 +1,5 @@
 import "./styles/GrantPage.css";
-import Header from "../../Header.js";
+import Header from "./Header";
 import GrantList from "./GrantList.js";
 import Footer from "./Footer.js";
 import BellButton from "../../Bell.js";
@@ -7,7 +7,7 @@ import "../../Bell.css";
 
 function GrantPage() {
   return (
-    <div className="grant-page">
+    <div className="grant-page p-5 bg-tan">
       <div className="top-half">
         <Header />
       </div>
