+++ frontend/src/grant-info/components/Header.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import "./styles/Header.css";
 import { useState } from "react";
-import logo from "./images/bcan_logo.svg";
+import logo from "../../images/bcan_logo.svg"
 
 interface NavBarProps {
   name: string;
@@ -25,14 +25,14 @@ const Header: React.FC = () => {
   const [selected, setSelected] = useState("All Grants");
 
   return (
-    <header className="header">
-      <header className="header-left-comp">
+    <header className="header mx-auto  max-w-8xl mb-5">
+      <header className="header-left-comp mx-5">
         <img className="logo" src={logo} alt="BCAN Logo" />
       </header>
-      <header className="header-right-comp">
-        <ul className="grant-buttons">
+      <header className="header-right-comp w-full">
+        <ul className="grant-buttons flex flex-row w-full justify-around">
           {linkList.map((item, index) => (
-            <li key={index}>
+            <li key={index} className="flex-grow w-auto">
               <Link
                 onClick={() => setSelected(item.name)}
                 style={{
