+++ frontend/src/grant-info/components/styles/GrantPage.css
@@ -12,6 +12,5 @@
 .bot-half {
     display: flex;
     flex-direction: column;
-    margin: 0 1rem;
 }
