--- frontend/src/styles/Header.css
@@ -1,46 +1,0 @@
-.header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    height: 100%; /* Fill up entire container */
-    margin: 2rem 2.5rem;
-    font-size: 1.2rem;
-}
-
-.header-right-comp {
-    display: flex;
-    width:100%;
-    justify-content: space-between;
-    align-items: center;
-    text-align: center;
-    margin-left: 20px;
-}
-
-.header-left-comp {
-    display: flex;
-    align-items: center;
-    list-style-type: none; /* this is a placeholder */
-}
-
-.grant-buttons {
-    display: flex;
-    width:100%;
-    margin: 0;
-    padding: 0;
-    color: black;
-    list-style-type: none;
-}
-
-.grant-buttons a {
-    color: black;
-}
-
-.grant-buttons li {
-    flex-basis: 20%;
-}
-
-
-.logo {
-    max-width: 100%;
-    height: 100px;
-}
